Created Youtube Player as component
