+++ client/pages/index.js
@@ -1,30 +1,15 @@
 import Head from 'next/head'
 import Image from 'next/image'
 import styles from '../styles/Home.module.css'
-import YouTube from 'react-youtube'
-import 'bootstrap/dist/css/bootstrap.min.css';
-import SearchBar from '../components/searchbar';
+
+// Components 
+import SearchBar from '../components/searchbar'
+import YoutubePlayer from '../components/YoutubePlayer'
 
 export default function Home() {
-
-  const onPlayerReady = (event) => {
-    // access to player in all event handlers via event.target
-    event.target.pauseVideo();
-  }
-
-  const opts = {
-    height: '390',
-    width: '640',
-    playerVars: {
-      // https://developers.google.com/youtube/player_parameters
-      autoplay: 1,
-    },
-  };
-
   return (
     <div className={styles.container}>
       <Head>
-    
         <title>Create Next App</title>
         <meta name="description" content="Generated by create next app" />
         <link rel="icon" href="/favicon.ico" />
@@ -39,7 +24,7 @@ export default function Home() {
           </nav>
         </header>
         <SearchBar/>
-        <YouTube videoId="2g811Eo7K8U" opts={opts} onReady={onPlayerReady} />
+        <YoutubePlayer/>
       </main>
 
     </div>
